Fix compiler bugs; add vm debugging
IIRC
- error on same-scope shadow
- fix &&
- arg list infinite loop

Also add poorly concieved test translated from a javascript game, just
like euler-3

+++ rvmfg/src/read.rs
@@ -147,7 +147,7 @@ pub fn read_fn_header(code: &[u8], mut ip: &mut usize, is_extern: bool) -> (Stri
     let codeloc = if is_extern {
         0
     } else {
-        read_u32(code, &mut ip)
+        read_u32(code, &mut ip) as usize
     };
     let func = Fn::new(codeloc, return_type, parameters);
     (name, func)
